+++ front/src/components/Reviews.js
@@ -1,11 +1,10 @@
 import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { getComment } from '../redux/ducks/comment';
-import { Comment, Divider, message } from 'antd';
+import { Comment, Divider } from 'antd';
 import { InfoWrapper } from '../globalStyles';
 import CommentForm from './CommentForm';
 import CommentList from './CommentList';
-import { commentService } from '../service/comments.js';
 
 const Reviews = () => {
   const dispatch = useDispatch();
@@ -17,17 +16,6 @@ const Reviews = () => {
   const comment = useSelector((state) => state.comment.comment);
   console.log('comment', comment);
 
-  const addComment = async (value) => {
-    try {
-      const res = await commentService.add(value);
-      console.log('post comment result : ', res);
-    } catch (e) {
-      console.log(e.message);
-    }
-    dispatch(getComment());
-    message.success('작성되었습니다.');
-  };
-
   return (
     <InfoWrapper>
       <h1>
@@ -36,7 +24,7 @@ const Reviews = () => {
           ({comment ? comment.length : 0})
         </span>
       </h1>
-      <Comment content={<CommentForm addComment={addComment} />} />
+      <Comment content={<CommentForm />} />
       <Divider style={{ marginTop: '-1rem' }} />
       {comment && <CommentList />}
     </InfoWrapper>
