+++ front/src/redux/sagas/requests/comment.js
@@ -1,5 +1,6 @@
 import axios from 'axios';
 
 export function requestGetComment() {
-  return axios.get('http://localhost:4000/comments');
+  // return axios.get(`http://localhost:4000/comments/${cafe_id}`);
+  return axios.get(`http://localhost:4000/comments`);
 }
